Added Doc strings to all methods and co olocated mapGoogleAddress to with other functions
