+++ checkout-ui-custom/checkout6-custom.js
@@ -266,7 +266,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
   \***********************************************/
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var _Elements_FormField__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./Elements/FormField */ \"./src/partials/Deliver/Elements/FormField.js\");\n/* harmony import */ var _utils__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./utils */ \"./src/partials/Deliver/utils.js\");\n\n\nvar populateAddressFromSearch = function populateAddressFromSearch(address) {\n  var _document$getElementB;\n  var street = address.street,\n    neighborhood = address.neighborhood,\n    postalCode = address.postalCode,\n    state = address.state,\n    city = address.city,\n    lat = address.lat,\n    lng = address.lng;\n\n  // Clear any populated fields\n  (_document$getElementB = document.getElementById('bash--address-form')) === null || _document$getElementB === void 0 ? void 0 : _document$getElementB.reset();\n\n  // Clear hidden ID fields to prevent overwriting existing.\n  document.getElementById('bash--input-addressId').value = '';\n  document.getElementById('bash--input-addressName').value = '';\n  document.getElementById('bash--input-number').value = '  ';\n  document.getElementById('bash--input-street').value = street !== null && street !== void 0 ? street : '';\n  document.getElementById('bash--input-neighborhood').value = neighborhood !== null && neighborhood !== void 0 ? neighborhood : '';\n  document.getElementById('bash--input-city').value = city !== null && city !== void 0 ? city : '';\n  document.getElementById('bash--input-postalCode').value = postalCode !== null && postalCode !== void 0 ? postalCode : '';\n  document.getElementById('bash--input-state').value = (0,_utils__WEBPACK_IMPORTED_MODULE_1__.provinceShortCode)(state);\n  document.getElementById('bash--input-lat').value = lat || '';\n  document.getElementById('bash--input-lng').value = lng || '';\n\n  // Update previously invalid fields.\n  $(':invalid').trigger('change');\n};\nvar checkForAddressResults = function checkForAddressResults(event) {\n  setTimeout(function () {\n    var _event$target;\n    var pacContainers = document.querySelectorAll('.pac-container');\n    var hiddenPacContainers = document.querySelectorAll(\".pac-container[style*='display: none']\");\n    if ((pacContainers === null || pacContainers === void 0 ? void 0 : pacContainers.length) === (hiddenPacContainers === null || hiddenPacContainers === void 0 ? void 0 : hiddenPacContainers.length) && ((_event$target = event.target) === null || _event$target === void 0 || (_event$target = _event$target.value) === null || _event$target === void 0 ? void 0 : _event$target.length) > 3) {\n      $('#address-search-field-container:not(.no-results)').addClass('no-results');\n    } else {\n      $('#address-search-field-container.no-results').removeClass('no-results');\n    }\n  }, 250);\n};\nvar initGoogleAutocomplete = function initGoogleAutocomplete() {\n  if (!window.google) return;\n  var input = document.getElementById('bash--input-address-search');\n  if (!input) return;\n  var autocomplete = new window.google.maps.places.Autocomplete(input, {\n    componentRestrictions: {\n      country: 'ZA'\n    }\n  });\n  window.google.maps.event.addListener(autocomplete, 'place_changed', function () {\n    var place = autocomplete.getPlace();\n    var addressComponents = place.address_components,\n      geometry = place.geometry;\n    var address = (0,_utils__WEBPACK_IMPORTED_MODULE_1__.mapGoogleAddress)(addressComponents, geometry);\n\n    // Populate the form\n    // Set view to add-address\n    populateAddressFromSearch(address);\n    window.postMessage({\n      action: 'setDeliveryView',\n      view: 'address-form'\n    });\n    input.value = '';\n  });\n  input === null || input === void 0 ? void 0 : input.addEventListener('keyup', checkForAddressResults);\n};\n\n/**\n * AddressNotFoundNotification - html template for address not found notification\n * @returns {string} - html string\n */\nvar AddressNotFoundNotification = function AddressNotFoundNotification() {\n  return (/* html */\"\\n  <div id=\\\"no-address-search-results-notification\\\" class=\\\"notification info\\\" >\\n    <span class=\\\"icon\\\"></span>\\n    <div class=\\\"notification-content\\\">\\n      We could not find your address. \\n      <a class=\\\"no-results-drop-down\\\" href=\\\"\\\" data-view=\\\"address-form\\\" id=\\\"no-address-search-results\\\">\\n        Please click here to enter it manually.\\n      </a>\\n    </div>\\n  </div>\\n\"\n  );\n};\n\n/**\n * AddressSearch - html template for initializaitng address search field and google autocomplete\n * @returns {string} - html string\n */\nvar AddressSearch = function AddressSearch() {\n  setTimeout(function () {\n    initGoogleAutocomplete();\n  }, 500);\n  var searchField = (0,_Elements_FormField__WEBPACK_IMPORTED_MODULE_0__[\"default\"])({\n    name: 'address-search',\n    placeholder: 'Start typing an address...',\n    autoComplete: 'off'\n  });\n  return (/* html */\"\\n    <div class=\\\"address-search-field-container\\\" id=\\\"address-search-field-container\\\">\\n      \".concat(searchField, \"\\n      \").concat(AddressNotFoundNotification(), \"\\n      <p>Wow Shaheen Large Refactor</p>\\n    </div>\\n  \")\n  );\n};\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (AddressSearch);\n\n//# sourceURL=webpack://custom-shipping-step-by-items/./src/partials/Deliver/AddressSearch.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var _Elements_FormField__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./Elements/FormField */ \"./src/partials/Deliver/Elements/FormField.js\");\n/* harmony import */ var _utils__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./utils */ \"./src/partials/Deliver/utils.js\");\nfunction _typeof(obj) { \"@babel/helpers - typeof\"; return _typeof = \"function\" == typeof Symbol && \"symbol\" == typeof Symbol.iterator ? function (obj) { return typeof obj; } : function (obj) { return obj && \"function\" == typeof Symbol && obj.constructor === Symbol && obj !== Symbol.prototype ? \"symbol\" : typeof obj; }, _typeof(obj); }\nfunction ownKeys(object, enumerableOnly) { var keys = Object.keys(object); if (Object.getOwnPropertySymbols) { var symbols = Object.getOwnPropertySymbols(object); enumerableOnly && (symbols = symbols.filter(function (sym) { return Object.getOwnPropertyDescriptor(object, sym).enumerable; })), keys.push.apply(keys, symbols); } return keys; }\nfunction _objectSpread(target) { for (var i = 1; i < arguments.length; i++) { var source = null != arguments[i] ? arguments[i] : {}; i % 2 ? ownKeys(Object(source), !0).forEach(function (key) { _defineProperty(target, key, source[key]); }) : Object.getOwnPropertyDescriptors ? Object.defineProperties(target, Object.getOwnPropertyDescriptors(source)) : ownKeys(Object(source)).forEach(function (key) { Object.defineProperty(target, key, Object.getOwnPropertyDescriptor(source, key)); }); } return target; }\nfunction _defineProperty(obj, key, value) { key = _toPropertyKey(key); if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }\nfunction _toPropertyKey(arg) { var key = _toPrimitive(arg, \"string\"); return _typeof(key) === \"symbol\" ? key : String(key); }\nfunction _toPrimitive(input, hint) { if (_typeof(input) !== \"object\" || input === null) return input; var prim = input[Symbol.toPrimitive]; if (prim !== undefined) { var res = prim.call(input, hint || \"default\"); if (_typeof(res) !== \"object\") return res; throw new TypeError(\"@@toPrimitive must return a primitive value.\"); } return (hint === \"string\" ? String : Number)(input); }\n\n\n\n/**\n * mapGoogleAddress - given a google address object, map it to a structured address object\n * @param {Array} addressComponents - google address components\n * @param {Object} geometry - google geometry object\n * @returns {Object} - structured address object\n */\nvar mapGoogleAddress = function mapGoogleAddress(addressComponents, geometry) {\n  var _addressComponents$fi, _addressComponents$fi2, _addressComponents$fi3, _addressComponents$fi4, _addressComponents$fi5, _addressComponents$fi6;\n  if (!addressComponents || addressComponents.length < 1) return {};\n  var streetNumber = (_addressComponents$fi = addressComponents.find(function (item) {\n    return item.types.includes('street_number');\n  })) === null || _addressComponents$fi === void 0 ? void 0 : _addressComponents$fi.long_name;\n  var street = (_addressComponents$fi2 = addressComponents.find(function (item) {\n    return item.types.includes('route');\n  })) === null || _addressComponents$fi2 === void 0 ? void 0 : _addressComponents$fi2.long_name;\n  var neighborhood = (_addressComponents$fi3 = addressComponents.find(function (item) {\n    return item.types.includes('sublocality');\n  })) === null || _addressComponents$fi3 === void 0 ? void 0 : _addressComponents$fi3.long_name;\n  var city = (_addressComponents$fi4 = addressComponents.find(function (item) {\n    return item.types.includes('locality');\n  })) === null || _addressComponents$fi4 === void 0 ? void 0 : _addressComponents$fi4.long_name;\n  var postalCode = (_addressComponents$fi5 = addressComponents.find(function (item) {\n    return item.types.includes('postal_code');\n  })) === null || _addressComponents$fi5 === void 0 ? void 0 : _addressComponents$fi5.long_name;\n  var state = (_addressComponents$fi6 = addressComponents.find(function (item) {\n    return item.types.includes('administrative_area_level_1');\n  })) === null || _addressComponents$fi6 === void 0 ? void 0 : _addressComponents$fi6.long_name;\n  var coords = {\n    lat: '',\n    lng: ''\n  };\n  if (geometry) {\n    coords.lat = geometry.location.lat();\n    coords.lng = geometry.location.lng();\n  }\n  return _objectSpread({\n    street: \"\".concat(streetNumber !== null && streetNumber !== void 0 ? streetNumber : '', \" \").concat(street !== null && street !== void 0 ? street : '').trim(),\n    neighborhood: neighborhood,\n    city: city,\n    postalCode: postalCode,\n    state: state\n  }, coords);\n};\n\n/**\n * populateAddressFromSearch - fills form fields based on the address returned from google auto complete\n * @param {Object} address\n * @returns {void}\n */\nvar populateAddressFromSearch = function populateAddressFromSearch(address) {\n  var _document$getElementB;\n  var street = address.street,\n    neighborhood = address.neighborhood,\n    postalCode = address.postalCode,\n    state = address.state,\n    city = address.city,\n    lat = address.lat,\n    lng = address.lng;\n\n  // Clear any populated fields\n  (_document$getElementB = document.getElementById('bash--address-form')) === null || _document$getElementB === void 0 ? void 0 : _document$getElementB.reset();\n\n  // Clear hidden ID fields to prevent overwriting existing.\n  document.getElementById('bash--input-addressId').value = '';\n  document.getElementById('bash--input-addressName').value = '';\n  document.getElementById('bash--input-number').value = '  ';\n  document.getElementById('bash--input-street').value = street !== null && street !== void 0 ? street : '';\n  document.getElementById('bash--input-neighborhood').value = neighborhood !== null && neighborhood !== void 0 ? neighborhood : '';\n  document.getElementById('bash--input-city').value = city !== null && city !== void 0 ? city : '';\n  document.getElementById('bash--input-postalCode').value = postalCode !== null && postalCode !== void 0 ? postalCode : '';\n  document.getElementById('bash--input-state').value = (0,_utils__WEBPACK_IMPORTED_MODULE_1__.provinceShortCode)(state);\n  document.getElementById('bash--input-lat').value = lat || '';\n  document.getElementById('bash--input-lng').value = lng || '';\n\n  // Update previously invalid fields.\n  $(':invalid').trigger('change');\n};\n\n/**\n * checkForAddressResults - checks if there are any address results to display a notification\n * @param {*} event\n * @returns {void}\n */\nvar checkForAddressResults = function checkForAddressResults(event) {\n  setTimeout(function () {\n    var _event$target;\n    var pacContainers = document.querySelectorAll('.pac-container');\n    var hiddenPacContainers = document.querySelectorAll(\".pac-container[style*='display: none']\");\n    if ((pacContainers === null || pacContainers === void 0 ? void 0 : pacContainers.length) === (hiddenPacContainers === null || hiddenPacContainers === void 0 ? void 0 : hiddenPacContainers.length) && ((_event$target = event.target) === null || _event$target === void 0 || (_event$target = _event$target.value) === null || _event$target === void 0 ? void 0 : _event$target.length) > 3) {\n      $('#address-search-field-container:not(.no-results)').addClass('no-results');\n    } else {\n      $('#address-search-field-container.no-results').removeClass('no-results');\n    }\n  }, 250);\n};\n\n/**\n * initGoogleAutocomplete - initializes google autocomplete on the address search field, add event listeners for place changed, change view on place changed\n * @returns {void}\n */\nvar initGoogleAutocomplete = function initGoogleAutocomplete() {\n  if (!window.google) return;\n  var input = document.getElementById('bash--input-address-search');\n  if (!input) return;\n  var autocomplete = new window.google.maps.places.Autocomplete(input, {\n    componentRestrictions: {\n      country: 'ZA'\n    }\n  });\n  window.google.maps.event.addListener(autocomplete, 'place_changed', function () {\n    var place = autocomplete.getPlace();\n    var addressComponents = place.address_components,\n      geometry = place.geometry;\n    var address = mapGoogleAddress(addressComponents, geometry);\n\n    // Populate the form\n    // Set view to add-address\n    populateAddressFromSearch(address);\n    window.postMessage({\n      action: 'setDeliveryView',\n      view: 'address-form'\n    });\n    input.value = '';\n  });\n  input === null || input === void 0 ? void 0 : input.addEventListener('keyup', checkForAddressResults);\n};\n\n/**\n * AddressNotFoundNotification - html template for address not found notification\n * @returns {string} - html string\n */\nvar AddressNotFoundNotification = function AddressNotFoundNotification() {\n  return (/* html */\"\\n  <div id=\\\"no-address-search-results-notification\\\" class=\\\"notification info\\\" >\\n    <span class=\\\"icon\\\"></span>\\n    <div class=\\\"notification-content\\\">\\n      We could not find your address. \\n      <a class=\\\"no-results-drop-down\\\" href=\\\"\\\" data-view=\\\"address-form\\\" id=\\\"no-address-search-results\\\">\\n        Please click here to enter it manually.\\n      </a>\\n    </div>\\n  </div>\\n\"\n  );\n};\n\n/**\n * AddressSearch - html template for initializaitng address search field and google autocomplete\n * @returns {string} - html string\n */\nvar AddressSearch = function AddressSearch() {\n  setTimeout(function () {\n    initGoogleAutocomplete();\n  }, 500);\n  var searchField = (0,_Elements_FormField__WEBPACK_IMPORTED_MODULE_0__[\"default\"])({\n    name: 'address-search',\n    placeholder: 'Start typing an address...',\n    autoComplete: 'off'\n  });\n  return (/* html */\"\\n    <div class=\\\"address-search-field-container\\\" id=\\\"address-search-field-container\\\">\\n      \".concat(searchField, \"\\n      \").concat(AddressNotFoundNotification(), \"\\n    </div>\\n  \")\n  );\n};\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (AddressSearch);\n\n//# sourceURL=webpack://custom-shipping-step-by-items/./src/partials/Deliver/AddressSearch.js?");
 
 /***/ }),
 
@@ -486,7 +486,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
   \***************************************/
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"addressIsValid\": () => (/* binding */ addressIsValid),\n/* harmony export */   \"clearRicaFields\": () => (/* binding */ clearRicaFields),\n/* harmony export */   \"customShippingDataIsValid\": () => (/* binding */ customShippingDataIsValid),\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__),\n/* harmony export */   \"getBestRecipient\": () => (/* binding */ getBestRecipient),\n/* harmony export */   \"mapGoogleAddress\": () => (/* binding */ mapGoogleAddress),\n/* harmony export */   \"parseAttribute\": () => (/* binding */ parseAttribute),\n/* harmony export */   \"populateAddressForm\": () => (/* binding */ populateAddressForm),\n/* harmony export */   \"populateDeliveryError\": () => (/* binding */ populateDeliveryError),\n/* harmony export */   \"populateExtraFields\": () => (/* binding */ populateExtraFields),\n/* harmony export */   \"populateRicaFields\": () => (/* binding */ populateRicaFields),\n/* harmony export */   \"populateTVFields\": () => (/* binding */ populateTVFields),\n/* harmony export */   \"provinceShortCode\": () => (/* binding */ provinceShortCode),\n/* harmony export */   \"setCartClasses\": () => (/* binding */ setCartClasses),\n/* harmony export */   \"setDeliveryLoading\": () => (/* binding */ setDeliveryLoading),\n/* harmony export */   \"setPickupLoading\": () => (/* binding */ setPickupLoading),\n/* harmony export */   \"showAlertBox\": () => (/* binding */ showAlertBox),\n/* harmony export */   \"updateDeliveryFeeDisplay\": () => (/* binding */ updateDeliveryFeeDisplay)\n/* harmony export */ });\n/* harmony import */ var _utils_const__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../../utils/const */ \"./src/utils/const.js\");\n/* harmony import */ var _utils_functions__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../../utils/functions */ \"./src/utils/functions.js\");\n/* harmony import */ var _utils_isInSouthAfrica__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../../utils/isInSouthAfrica */ \"./src/utils/isInSouthAfrica.js\");\n/* harmony import */ var _utils_phoneFields__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ../../utils/phoneFields */ \"./src/utils/phoneFields.js\");\n/* harmony import */ var _utils_services__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ../../utils/services */ \"./src/utils/services.js\");\n/* harmony import */ var _DeliveryError__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! ./DeliveryError */ \"./src/partials/Deliver/DeliveryError.js\");\n/* harmony import */ var _Elements_Alert__WEBPACK_IMPORTED_MODULE_6__ = __webpack_require__(/*! ./Elements/Alert */ \"./src/partials/Deliver/Elements/Alert.js\");\n/* harmony import */ var _constants__WEBPACK_IMPORTED_MODULE_7__ = __webpack_require__(/*! ./constants */ \"./src/partials/Deliver/constants.js\");\n/* harmony import */ var _utils_phoneNumberFormat__WEBPACK_IMPORTED_MODULE_8__ = __webpack_require__(/*! ../../utils/phoneNumberFormat */ \"./src/utils/phoneNumberFormat.js\");\nfunction _typeof(obj) { \"@babel/helpers - typeof\"; return _typeof = \"function\" == typeof Symbol && \"symbol\" == typeof Symbol.iterator ? function (obj) { return typeof obj; } : function (obj) { return obj && \"function\" == typeof Symbol && obj.constructor === Symbol && obj !== Symbol.prototype ? \"symbol\" : typeof obj; }, _typeof(obj); }\nfunction _toConsumableArray(arr) { return _arrayWithoutHoles(arr) || _iterableToArray(arr) || _unsupportedIterableToArray(arr) || _nonIterableSpread(); }\nfunction _nonIterableSpread() { throw new TypeError(\"Invalid attempt to spread non-iterable instance.\\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.\"); }\nfunction _iterableToArray(iter) { if (typeof Symbol !== \"undefined\" && iter[Symbol.iterator] != null || iter[\"@@iterator\"] != null) return Array.from(iter); }\nfunction _arrayWithoutHoles(arr) { if (Array.isArray(arr)) return _arrayLikeToArray(arr); }\nfunction _regeneratorRuntime() { \"use strict\"; /*! regenerator-runtime -- Copyright (c) 2014-present, Facebook, Inc. -- license (MIT): https://github.com/facebook/regenerator/blob/main/LICENSE */ _regeneratorRuntime = function _regeneratorRuntime() { return exports; }; var exports = {}, Op = Object.prototype, hasOwn = Op.hasOwnProperty, defineProperty = Object.defineProperty || function (obj, key, desc) { obj[key] = desc.value; }, $Symbol = \"function\" == typeof Symbol ? Symbol : {}, iteratorSymbol = $Symbol.iterator || \"@@iterator\", asyncIteratorSymbol = $Symbol.asyncIterator || \"@@asyncIterator\", toStringTagSymbol = $Symbol.toStringTag || \"@@toStringTag\"; function define(obj, key, value) { return Object.defineProperty(obj, key, { value: value, enumerable: !0, configurable: !0, writable: !0 }), obj[key]; } try { define({}, \"\"); } catch (err) { define = function define(obj, key, value) { return obj[key] = value; }; } function wrap(innerFn, outerFn, self, tryLocsList) { var protoGenerator = outerFn && outerFn.prototype instanceof Generator ? outerFn : Generator, generator = Object.create(protoGenerator.prototype), context = new Context(tryLocsList || []); return defineProperty(generator, \"_invoke\", { value: makeInvokeMethod(innerFn, self, context) }), generator; } function tryCatch(fn, obj, arg) { try { return { type: \"normal\", arg: fn.call(obj, arg) }; } catch (err) { return { type: \"throw\", arg: err }; } } exports.wrap = wrap; var ContinueSentinel = {}; function Generator() {} function GeneratorFunction() {} function GeneratorFunctionPrototype() {} var IteratorPrototype = {}; define(IteratorPrototype, iteratorSymbol, function () { return this; }); var getProto = Object.getPrototypeOf, NativeIteratorPrototype = getProto && getProto(getProto(values([]))); NativeIteratorPrototype && NativeIteratorPrototype !== Op && hasOwn.call(NativeIteratorPrototype, iteratorSymbol) && (IteratorPrototype = NativeIteratorPrototype); var Gp = GeneratorFunctionPrototype.prototype = Generator.prototype = Object.create(IteratorPrototype); function defineIteratorMethods(prototype) { [\"next\", \"throw\", \"return\"].forEach(function (method) { define(prototype, method, function (arg) { return this._invoke(method, arg); }); }); } function AsyncIterator(generator, PromiseImpl) { function invoke(method, arg, resolve, reject) { var record = tryCatch(generator[method], generator, arg); if (\"throw\" !== record.type) { var result = record.arg, value = result.value; return value && \"object\" == _typeof(value) && hasOwn.call(value, \"__await\") ? PromiseImpl.resolve(value.__await).then(function (value) { invoke(\"next\", value, resolve, reject); }, function (err) { invoke(\"throw\", err, resolve, reject); }) : PromiseImpl.resolve(value).then(function (unwrapped) { result.value = unwrapped, resolve(result); }, function (error) { return invoke(\"throw\", error, resolve, reject); }); } reject(record.arg); } var previousPromise; defineProperty(this, \"_invoke\", { value: function value(method, arg) { function callInvokeWithMethodAndArg() { return new PromiseImpl(function (resolve, reject) { invoke(method, arg, resolve, reject); }); } return previousPromise = previousPromise ? previousPromise.then(callInvokeWithMethodAndArg, callInvokeWithMethodAndArg) : callInvokeWithMethodAndArg(); } }); } function makeInvokeMethod(innerFn, self, context) { var state = \"suspendedStart\"; return function (method, arg) { if (\"executing\" === state) throw new Error(\"Generator is already running\"); if (\"completed\" === state) { if (\"throw\" === method) throw arg; return { value: void 0, done: !0 }; } for (context.method = method, context.arg = arg;;) { var delegate = context.delegate; if (delegate) { var delegateResult = maybeInvokeDelegate(delegate, context); if (delegateResult) { if (delegateResult === ContinueSentinel) continue; return delegateResult; } } if (\"next\" === context.method) context.sent = context._sent = context.arg;else if (\"throw\" === context.method) { if (\"suspendedStart\" === state) throw state = \"completed\", context.arg; context.dispatchException(context.arg); } else \"return\" === context.method && context.abrupt(\"return\", context.arg); state = \"executing\"; var record = tryCatch(innerFn, self, context); if (\"normal\" === record.type) { if (state = context.done ? \"completed\" : \"suspendedYield\", record.arg === ContinueSentinel) continue; return { value: record.arg, done: context.done }; } \"throw\" === record.type && (state = \"completed\", context.method = \"throw\", context.arg = record.arg); } }; } function maybeInvokeDelegate(delegate, context) { var methodName = context.method, method = delegate.iterator[methodName]; if (undefined === method) return context.delegate = null, \"throw\" === methodName && delegate.iterator[\"return\"] && (context.method = \"return\", context.arg = undefined, maybeInvokeDelegate(delegate, context), \"throw\" === context.method) || \"return\" !== methodName && (context.method = \"throw\", context.arg = new TypeError(\"The iterator does not provide a '\" + methodName + \"' method\")), ContinueSentinel; var record = tryCatch(method, delegate.iterator, context.arg); if (\"throw\" === record.type) return context.method = \"throw\", context.arg = record.arg, context.delegate = null, ContinueSentinel; var info = record.arg; return info ? info.done ? (context[delegate.resultName] = info.value, context.next = delegate.nextLoc, \"return\" !== context.method && (context.method = \"next\", context.arg = undefined), context.delegate = null, ContinueSentinel) : info : (context.method = \"throw\", context.arg = new TypeError(\"iterator result is not an object\"), context.delegate = null, ContinueSentinel); } function pushTryEntry(locs) { var entry = { tryLoc: locs[0] }; 1 in locs && (entry.catchLoc = locs[1]), 2 in locs && (entry.finallyLoc = locs[2], entry.afterLoc = locs[3]), this.tryEntries.push(entry); } function resetTryEntry(entry) { var record = entry.completion || {}; record.type = \"normal\", delete record.arg, entry.completion = record; } function Context(tryLocsList) { this.tryEntries = [{ tryLoc: \"root\" }], tryLocsList.forEach(pushTryEntry, this), this.reset(!0); } function values(iterable) { if (iterable || \"\" === iterable) { var iteratorMethod = iterable[iteratorSymbol]; if (iteratorMethod) return iteratorMethod.call(iterable); if (\"function\" == typeof iterable.next) return iterable; if (!isNaN(iterable.length)) { var i = -1, next = function next() { for (; ++i < iterable.length;) if (hasOwn.call(iterable, i)) return next.value = iterable[i], next.done = !1, next; return next.value = undefined, next.done = !0, next; }; return next.next = next; } } throw new TypeError(_typeof(iterable) + \" is not iterable\"); } return GeneratorFunction.prototype = GeneratorFunctionPrototype, defineProperty(Gp, \"constructor\", { value: GeneratorFunctionPrototype, configurable: !0 }), defineProperty(GeneratorFunctionPrototype, \"constructor\", { value: GeneratorFunction, configurable: !0 }), GeneratorFunction.displayName = define(GeneratorFunctionPrototype, toStringTagSymbol, \"GeneratorFunction\"), exports.isGeneratorFunction = function (genFun) { var ctor = \"function\" == typeof genFun && genFun.constructor; return !!ctor && (ctor === GeneratorFunction || \"GeneratorFunction\" === (ctor.displayName || ctor.name)); }, exports.mark = function (genFun) { return Object.setPrototypeOf ? Object.setPrototypeOf(genFun, GeneratorFunctionPrototype) : (genFun.__proto__ = GeneratorFunctionPrototype, define(genFun, toStringTagSymbol, \"GeneratorFunction\")), genFun.prototype = Object.create(Gp), genFun; }, exports.awrap = function (arg) { return { __await: arg }; }, defineIteratorMethods(AsyncIterator.prototype), define(AsyncIterator.prototype, asyncIteratorSymbol, function () { return this; }), exports.AsyncIterator = AsyncIterator, exports.async = function (innerFn, outerFn, self, tryLocsList, PromiseImpl) { void 0 === PromiseImpl && (PromiseImpl = Promise); var iter = new AsyncIterator(wrap(innerFn, outerFn, self, tryLocsList), PromiseImpl); return exports.isGeneratorFunction(outerFn) ? iter : iter.next().then(function (result) { return result.done ? result.value : iter.next(); }); }, defineIteratorMethods(Gp), define(Gp, toStringTagSymbol, \"Generator\"), define(Gp, iteratorSymbol, function () { return this; }), define(Gp, \"toString\", function () { return \"[object Generator]\"; }), exports.keys = function (val) { var object = Object(val), keys = []; for (var key in object) keys.push(key); return keys.reverse(), function next() { for (; keys.length;) { var key = keys.pop(); if (key in object) return next.value = key, next.done = !1, next; } return next.done = !0, next; }; }, exports.values = values, Context.prototype = { constructor: Context, reset: function reset(skipTempReset) { if (this.prev = 0, this.next = 0, this.sent = this._sent = undefined, this.done = !1, this.delegate = null, this.method = \"next\", this.arg = undefined, this.tryEntries.forEach(resetTryEntry), !skipTempReset) for (var name in this) \"t\" === name.charAt(0) && hasOwn.call(this, name) && !isNaN(+name.slice(1)) && (this[name] = undefined); }, stop: function stop() { this.done = !0; var rootRecord = this.tryEntries[0].completion; if (\"throw\" === rootRecord.type) throw rootRecord.arg; return this.rval; }, dispatchException: function dispatchException(exception) { if (this.done) throw exception; var context = this; function handle(loc, caught) { return record.type = \"throw\", record.arg = exception, context.next = loc, caught && (context.method = \"next\", context.arg = undefined), !!caught; } for (var i = this.tryEntries.length - 1; i >= 0; --i) { var entry = this.tryEntries[i], record = entry.completion; if (\"root\" === entry.tryLoc) return handle(\"end\"); if (entry.tryLoc <= this.prev) { var hasCatch = hasOwn.call(entry, \"catchLoc\"), hasFinally = hasOwn.call(entry, \"finallyLoc\"); if (hasCatch && hasFinally) { if (this.prev < entry.catchLoc) return handle(entry.catchLoc, !0); if (this.prev < entry.finallyLoc) return handle(entry.finallyLoc); } else if (hasCatch) { if (this.prev < entry.catchLoc) return handle(entry.catchLoc, !0); } else { if (!hasFinally) throw new Error(\"try statement without catch or finally\"); if (this.prev < entry.finallyLoc) return handle(entry.finallyLoc); } } } }, abrupt: function abrupt(type, arg) { for (var i = this.tryEntries.length - 1; i >= 0; --i) { var entry = this.tryEntries[i]; if (entry.tryLoc <= this.prev && hasOwn.call(entry, \"finallyLoc\") && this.prev < entry.finallyLoc) { var finallyEntry = entry; break; } } finallyEntry && (\"break\" === type || \"continue\" === type) && finallyEntry.tryLoc <= arg && arg <= finallyEntry.finallyLoc && (finallyEntry = null); var record = finallyEntry ? finallyEntry.completion : {}; return record.type = type, record.arg = arg, finallyEntry ? (this.method = \"next\", this.next = finallyEntry.finallyLoc, ContinueSentinel) : this.complete(record); }, complete: function complete(record, afterLoc) { if (\"throw\" === record.type) throw record.arg; return \"break\" === record.type || \"continue\" === record.type ? this.next = record.arg : \"return\" === record.type ? (this.rval = this.arg = record.arg, this.method = \"return\", this.next = \"end\") : \"normal\" === record.type && afterLoc && (this.next = afterLoc), ContinueSentinel; }, finish: function finish(finallyLoc) { for (var i = this.tryEntries.length - 1; i >= 0; --i) { var entry = this.tryEntries[i]; if (entry.finallyLoc === finallyLoc) return this.complete(entry.completion, entry.afterLoc), resetTryEntry(entry), ContinueSentinel; } }, \"catch\": function _catch(tryLoc) { for (var i = this.tryEntries.length - 1; i >= 0; --i) { var entry = this.tryEntries[i]; if (entry.tryLoc === tryLoc) { var record = entry.completion; if (\"throw\" === record.type) { var thrown = record.arg; resetTryEntry(entry); } return thrown; } } throw new Error(\"illegal catch attempt\"); }, delegateYield: function delegateYield(iterable, resultName, nextLoc) { return this.delegate = { iterator: values(iterable), resultName: resultName, nextLoc: nextLoc }, \"next\" === this.method && (this.arg = undefined), ContinueSentinel; } }, exports; }\nfunction asyncGeneratorStep(gen, resolve, reject, _next, _throw, key, arg) { try { var info = gen[key](arg); var value = info.value; } catch (error) { reject(error); return; } if (info.done) { resolve(value); } else { Promise.resolve(value).then(_next, _throw); } }\nfunction _asyncToGenerator(fn) { return function () { var self = this, args = arguments; return new Promise(function (resolve, reject) { var gen = fn.apply(self, args); function _next(value) { asyncGeneratorStep(gen, resolve, reject, _next, _throw, \"next\", value); } function _throw(err) { asyncGeneratorStep(gen, resolve, reject, _next, _throw, \"throw\", err); } _next(undefined); }); }; }\nfunction _slicedToArray(arr, i) { return _arrayWithHoles(arr) || _iterableToArrayLimit(arr, i) || _unsupportedIterableToArray(arr, i) || _nonIterableRest(); }\nfunction _nonIterableRest() { throw new TypeError(\"Invalid attempt to destructure non-iterable instance.\\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.\"); }\nfunction _unsupportedIterableToArray(o, minLen) { if (!o) return; if (typeof o === \"string\") return _arrayLikeToArray(o, minLen); var n = Object.prototype.toString.call(o).slice(8, -1); if (n === \"Object\" && o.constructor) n = o.constructor.name; if (n === \"Map\" || n === \"Set\") return Array.from(o); if (n === \"Arguments\" || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)) return _arrayLikeToArray(o, minLen); }\nfunction _arrayLikeToArray(arr, len) { if (len == null || len > arr.length) len = arr.length; for (var i = 0, arr2 = new Array(len); i < len; i++) arr2[i] = arr[i]; return arr2; }\nfunction _iterableToArrayLimit(arr, i) { var _i = null == arr ? null : \"undefined\" != typeof Symbol && arr[Symbol.iterator] || arr[\"@@iterator\"]; if (null != _i) { var _s, _e, _x, _r, _arr = [], _n = !0, _d = !1; try { if (_x = (_i = _i.call(arr)).next, 0 === i) { if (Object(_i) !== _i) return; _n = !1; } else for (; !(_n = (_s = _x.call(_i)).done) && (_arr.push(_s.value), _arr.length !== i); _n = !0); } catch (err) { _d = !0, _e = err; } finally { try { if (!_n && null != _i[\"return\"] && (_r = _i[\"return\"](), Object(_r) !== _r)) return; } finally { if (_d) throw _e; } } return _arr; } }\nfunction _arrayWithHoles(arr) { if (Array.isArray(arr)) return arr; }\nfunction ownKeys(object, enumerableOnly) { var keys = Object.keys(object); if (Object.getOwnPropertySymbols) { var symbols = Object.getOwnPropertySymbols(object); enumerableOnly && (symbols = symbols.filter(function (sym) { return Object.getOwnPropertyDescriptor(object, sym).enumerable; })), keys.push.apply(keys, symbols); } return keys; }\nfunction _objectSpread(target) { for (var i = 1; i < arguments.length; i++) { var source = null != arguments[i] ? arguments[i] : {}; i % 2 ? ownKeys(Object(source), !0).forEach(function (key) { _defineProperty(target, key, source[key]); }) : Object.getOwnPropertyDescriptors ? Object.defineProperties(target, Object.getOwnPropertyDescriptors(source)) : ownKeys(Object(source)).forEach(function (key) { Object.defineProperty(target, key, Object.getOwnPropertyDescriptor(source, key)); }); } return target; }\nfunction _defineProperty(obj, key, value) { key = _toPropertyKey(key); if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }\nfunction _toPropertyKey(arg) { var key = _toPrimitive(arg, \"string\"); return _typeof(key) === \"symbol\" ? key : String(key); }\nfunction _toPrimitive(input, hint) { if (_typeof(input) !== \"object\" || input === null) return input; var prim = input[Symbol.toPrimitive]; if (prim !== undefined) { var res = prim.call(input, hint || \"default\"); if (_typeof(res) !== \"object\") return res; throw new TypeError(\"@@toPrimitive must return a primitive value.\"); } return (hint === \"string\" ? String : Number)(input); }\n// @ts-nocheck\n\n\n\n\n\n\n\n\n\nvar _usePhoneNumberFormat = (0,_utils_phoneNumberFormat__WEBPACK_IMPORTED_MODULE_8__[\"default\"])(),\n  formatPhoneNumber = _usePhoneNumberFormat.formatPhoneNumber,\n  isValidNumber = _usePhoneNumberFormat.isValidNumber;\nvar setDeliveryLoading = function setDeliveryLoading() {\n  var _document$querySelect;\n  (_document$querySelect = document.querySelector('.bash--delivery-container')) === null || _document$querySelect === void 0 ? void 0 : _document$querySelect.classList.add('shimmer');\n};\nvar setPickupLoading = function setPickupLoading() {\n  var _document$querySelect2, _document$querySelect3;\n  (_document$querySelect2 = document.querySelector('.delivery-group-content')) === null || _document$querySelect2 === void 0 || (_document$querySelect2 = _document$querySelect2.classList) === null || _document$querySelect2 === void 0 ? void 0 : _document$querySelect2.add('shimmer');\n  (_document$querySelect3 = document.querySelector('.vtex-omnishipping-1-x-ask')) === null || _document$querySelect3 === void 0 || (_document$querySelect3 = _document$querySelect3.classList) === null || _document$querySelect3 === void 0 ? void 0 : _document$querySelect3.add('shimmer');\n};\nvar mapGoogleAddress = function mapGoogleAddress(addressComponents, geometry) {\n  var _addressComponents$fi, _addressComponents$fi2, _addressComponents$fi3, _addressComponents$fi4, _addressComponents$fi5, _addressComponents$fi6;\n  if (!addressComponents || addressComponents.length < 1) return {};\n  var streetNumber = (_addressComponents$fi = addressComponents.find(function (item) {\n    return item.types.includes('street_number');\n  })) === null || _addressComponents$fi === void 0 ? void 0 : _addressComponents$fi.long_name;\n  var street = (_addressComponents$fi2 = addressComponents.find(function (item) {\n    return item.types.includes('route');\n  })) === null || _addressComponents$fi2 === void 0 ? void 0 : _addressComponents$fi2.long_name;\n  var neighborhood = (_addressComponents$fi3 = addressComponents.find(function (item) {\n    return item.types.includes('sublocality');\n  })) === null || _addressComponents$fi3 === void 0 ? void 0 : _addressComponents$fi3.long_name;\n  var city = (_addressComponents$fi4 = addressComponents.find(function (item) {\n    return item.types.includes('locality');\n  })) === null || _addressComponents$fi4 === void 0 ? void 0 : _addressComponents$fi4.long_name;\n  var postalCode = (_addressComponents$fi5 = addressComponents.find(function (item) {\n    return item.types.includes('postal_code');\n  })) === null || _addressComponents$fi5 === void 0 ? void 0 : _addressComponents$fi5.long_name;\n  var state = (_addressComponents$fi6 = addressComponents.find(function (item) {\n    return item.types.includes('administrative_area_level_1');\n  })) === null || _addressComponents$fi6 === void 0 ? void 0 : _addressComponents$fi6.long_name;\n  var coords = {\n    lat: '',\n    lng: ''\n  };\n  if (geometry) {\n    coords.lat = geometry.location.lat();\n    coords.lng = geometry.location.lng();\n  }\n  return _objectSpread({\n    street: \"\".concat(streetNumber !== null && streetNumber !== void 0 ? streetNumber : '', \" \").concat(street !== null && street !== void 0 ? street : '').trim(),\n    neighborhood: neighborhood,\n    city: city,\n    postalCode: postalCode,\n    state: state\n  }, coords);\n};\nvar provinceShortCode = function provinceShortCode(province) {\n  switch (province) {\n    case 'Select':\n      return '';\n    case 'Western Cape':\n      return 'WC';\n    case 'Easter Cape':\n      return 'EC';\n    case 'Gauteng':\n      return 'GP';\n    case 'KwaZulu-Natal':\n    case 'KwaZulu Natal':\n      return 'KZN';\n    case 'Northern Cape':\n      return 'NC';\n    case 'Limpopo':\n      return 'LP';\n    case 'Mpumalanga':\n      return 'MP';\n    case 'North West':\n      return 'NW';\n    case 'Freestate':\n    case 'Free State':\n      return 'FS';\n    default:\n      return province;\n  }\n};\nvar getBestRecipient = function getBestRecipient(_ref) {\n  var _window, _window2, _window3, _document$getElementB;\n  var _ref$preferred = _ref.preferred,\n    preferred = _ref$preferred === void 0 ? undefined : _ref$preferred,\n    _ref$type = _ref.type,\n    type = _ref$type === void 0 ? 'delivery' : _ref$type;\n  var firstName = (_window = window) === null || _window === void 0 || (_window = _window.vtexjs) === null || _window === void 0 || (_window = _window.checkout) === null || _window === void 0 || (_window = _window.orderForm) === null || _window === void 0 || (_window = _window.clientProfileData) === null || _window === void 0 ? void 0 : _window.firstName;\n  var lastName = (_window2 = window) === null || _window2 === void 0 || (_window2 = _window2.vtexjs) === null || _window2 === void 0 || (_window2 = _window2.checkout) === null || _window2 === void 0 || (_window2 = _window2.orderForm) === null || _window2 === void 0 || (_window2 = _window2.clientProfileData) === null || _window2 === void 0 ? void 0 : _window2.lastName;\n  var shippingReceiverName = (_window3 = window) === null || _window3 === void 0 || (_window3 = _window3.vtexjs) === null || _window3 === void 0 || (_window3 = _window3.checkout) === null || _window3 === void 0 || (_window3 = _window3.orderForm) === null || _window3 === void 0 || (_window3 = _window3.shippingData) === null || _window3 === void 0 || (_window3 = _window3.address) === null || _window3 === void 0 ? void 0 : _window3.receiverName;\n  var clientProfileName = \"\".concat(firstName !== null && firstName !== void 0 ? firstName : '', \" \").concat(lastName !== null && lastName !== void 0 ? lastName : '').trim();\n  if (type === 'collect') return preferred || shippingReceiverName || clientProfileName || '';\n  return preferred || ((_document$getElementB = document.getElementById('client-first-name')) === null || _document$getElementB === void 0 ? void 0 : _document$getElementB.value) || clientProfileName || '';\n};\nvar populateAddressForm = function populateAddressForm(address) {\n  var _document$getElementB2, _document$getElementB3;\n  var number = address.number,\n    street = address.street,\n    addressType = address.addressType,\n    businessName = address.businessName,\n    companyBuilding = address.companyBuilding,\n    neighborhood = address.neighborhood,\n    postalCode = address.postalCode,\n    state = address.state,\n    city = address.city,\n    receiverName = address.receiverName,\n    receiverPhone = address.receiverPhone,\n    complement = address.complement,\n    id = address.id,\n    addressId = address.addressId,\n    addressName = address.addressName,\n    geoCoordinate = address.geoCoordinate;\n\n  // Clear any populated fields\n  (_document$getElementB2 = document.getElementById('bash--address-form')) === null || _document$getElementB2 === void 0 ? void 0 : _document$getElementB2.reset();\n  (0,_utils_functions__WEBPACK_IMPORTED_MODULE_1__.hideBusinessName)();\n  var lat;\n  var lng;\n  try {\n    var _JSON$parse = JSON.parse(JSON.stringify(geoCoordinate));\n    var _JSON$parse2 = _slicedToArray(_JSON$parse, 2);\n    lng = _JSON$parse2[0];\n    lat = _JSON$parse2[1];\n    if (!(0,_utils_isInSouthAfrica__WEBPACK_IMPORTED_MODULE_2__[\"default\"])([lng, lat])) {\n      // Coordinates are most likely still at [lat, lng]\n      if (_utils_isInSouthAfrica__WEBPACK_IMPORTED_MODULE_2__[\"default\"][(lat, lng)]) {\n        lng = lat;\n        lat = lng;\n      } else {\n        // Unset the coordinates\n        lng = 0;\n        lat = 0;\n      }\n    }\n  } catch (e) {\n    console.warn('Could not parse geo coords', {\n      address: address,\n      geoCoordinate: geoCoordinate\n    });\n  }\n\n  // Only overwrite defaults if values exist.\n  if (receiverName) document.getElementById('bash--input-receiverName').value = receiverName !== null && receiverName !== void 0 ? receiverName : '';\n  if (complement) document.getElementById('bash--input-complement').value = complement !== null && complement !== void 0 ? complement : '';\n\n  // addressId indicates that address is being edited / completed.\n  if (id || addressId) document.getElementById('bash--input-addressId').value = id || addressId; // TODO remove this?\n  if (addressName) document.getElementById('bash--input-addressName').value = addressName;\n  var streetLine = \"\".concat(number ? \"\".concat(number, \" \") : '').concat(street).replace(\", \".concat(companyBuilding), '');\n\n  // Address type\n  if (addressType === 'commercial') {\n    $('#radio-addressType-business').click();\n    (0,_utils_functions__WEBPACK_IMPORTED_MODULE_1__.showBusinessName)({\n      focus: false\n    });\n  }\n  if (businessName) document.getElementById('bash--input-businessName').value = businessName;\n  document.getElementById('bash--input-number').value = '';\n  document.getElementById('bash--input-street').value = streetLine || '';\n  document.getElementById('bash--input-companyBuilding').value = companyBuilding || '';\n  document.getElementById('bash--input-neighborhood').value = neighborhood || '';\n  document.getElementById('bash--input-city').value = city || '';\n  document.getElementById('bash--input-postalCode').value = postalCode || '';\n  document.getElementById('bash--input-state').value = provinceShortCode(state);\n  document.getElementById('bash--input-lat').value = lat || '';\n  document.getElementById('bash--input-lng').value = lng || '';\n  var fields = (0,_utils_services__WEBPACK_IMPORTED_MODULE_4__.getOrderFormCustomData)(_utils_const__WEBPACK_IMPORTED_MODULE_0__.PICKUP);\n\n  // Only overwrite defaults if values exist.\n  if (receiverName) document.getElementById('bash--input-receiverName').value = receiverName !== null && receiverName !== void 0 ? receiverName : '';\n  if (complement) document.getElementById('bash--input-complement').value = complement !== null && complement !== void 0 ? complement : '';\n  document.getElementById('bash--input-receiverPhone').value = formatPhoneNumber(receiverPhone).trim() || (0,_utils_phoneFields__WEBPACK_IMPORTED_MODULE_3__.getBestPhoneNumber)({\n    preferred: receiverPhone,\n    type: 'delivery',\n    fields: fields\n  });\n  (_document$getElementB3 = document.getElementById('bash--input-receiverPhone')) === null || _document$getElementB3 === void 0 ? void 0 : _document$getElementB3.addEventListener('onblur', function (event) {\n    event.target.value = event.target.value.trim();\n  });\n  $(':invalid').trigger('change');\n};\nvar parseAttribute = function parseAttribute(data) {\n  try {\n    return JSON.parse(decodeURIComponent(data));\n  } catch (e) {\n    return undefined;\n  }\n};\nvar populateExtraFields = function populateExtraFields(address, fields) {\n  var prefix = arguments.length > 2 && arguments[2] !== undefined ? arguments[2] : '';\n  var override = arguments.length > 3 && arguments[3] !== undefined ? arguments[3] : false;\n  if (!address) return;\n  for (var i = 0; i < fields.length; i++) {\n    var fieldId = \"bash--input-\".concat(prefix).concat(fields[i]);\n    if (document.getElementById(fieldId) && (address[fields[i]] || override) && (!document.getElementById(fieldId).value || override)) {\n      document.getElementById(fieldId).value = address[fields[i]];\n    }\n  }\n  $(':invalid').trigger('change');\n};\nvar populateRicaFields = function populateRicaFields() {\n  var _document$getElementB4;\n  var address = window.vtexjs.checkout.orderForm.shippingData.address;\n  if ((_document$getElementB4 = document.getElementById('bash--input-rica_streetAddress')) !== null && _document$getElementB4 !== void 0 && _document$getElementB4.value || !address) return;\n  address.fullName = getBestRecipient({\n    type: 'delivery'\n  });\n  address.streetAddress = address.street;\n  address.suburb = address.neighborhood;\n  address.province = address.state;\n  populateExtraFields(address, _constants__WEBPACK_IMPORTED_MODULE_7__.requiredRicaFields, 'rica_');\n  var data = (0,_utils_services__WEBPACK_IMPORTED_MODULE_4__.getOrderFormCustomData)(_utils_const__WEBPACK_IMPORTED_MODULE_0__.RICA_APP);\n  if (data.streetAddress) populateExtraFields(data, _constants__WEBPACK_IMPORTED_MODULE_7__.requiredRicaFields, 'rica_', true);\n};\nvar clearRicaFields = function clearRicaFields() {\n  var idOrPassport = $('#bash--input-rica_idOrPassport').val();\n  var clearedRica = {\n    idOrPassport: idOrPassport !== null && idOrPassport !== void 0 ? idOrPassport : '',\n    // TODO populate with users ID\n    fullName: '',\n    streetAddress: '',\n    suburb: '',\n    city: '',\n    postalCode: '',\n    province: ''\n  };\n  populateExtraFields(clearedRica, _constants__WEBPACK_IMPORTED_MODULE_7__.requiredRicaFields, 'rica_', true);\n};\nvar populateTVFields = /*#__PURE__*/function () {\n  var _ref2 = _asyncToGenerator( /*#__PURE__*/_regeneratorRuntime().mark(function _callee() {\n    var data;\n    return _regeneratorRuntime().wrap(function _callee$(_context) {\n      while (1) switch (_context.prev = _context.next) {\n        case 0:\n          data = (0,_utils_services__WEBPACK_IMPORTED_MODULE_4__.getOrderFormCustomData)(_utils_const__WEBPACK_IMPORTED_MODULE_0__.TV_APP);\n          populateExtraFields(data, _constants__WEBPACK_IMPORTED_MODULE_7__.requiredTVFields, 'tv');\n        case 2:\n        case \"end\":\n          return _context.stop();\n      }\n    }, _callee);\n  }));\n  return function populateTVFields() {\n    return _ref2.apply(this, arguments);\n  };\n}();\n\n// Runs when you setAddress\nvar addressIsValid = function addressIsValid(address) {\n  var validateExtraFields = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : true;\n  var items = window.vtexjs.checkout.orderForm.items;\n  var _getSpecialCategories = (0,_utils_functions__WEBPACK_IMPORTED_MODULE_1__.getSpecialCategories)(items),\n    hasTVs = _getSpecialCategories.hasTVs,\n    hasSimCards = _getSpecialCategories.hasSimCards;\n  var requiredFields = _constants__WEBPACK_IMPORTED_MODULE_7__.requiredAddressFields;\n  var invalidFields = [];\n  if (hasTVs && validateExtraFields) {\n    requiredFields = [].concat(_toConsumableArray(requiredFields), _toConsumableArray(_constants__WEBPACK_IMPORTED_MODULE_7__.requiredTVFields));\n  }\n  if (hasSimCards && validateExtraFields) {\n    requiredFields = [].concat(_toConsumableArray(requiredFields), _toConsumableArray(_constants__WEBPACK_IMPORTED_MODULE_7__.requiredRicaFields));\n  }\n  for (var i = 0; i < requiredFields.length; i++) {\n    if (!address[requiredFields[i]]) invalidFields.push(requiredFields[i]);\n  }\n  if (requiredFields.includes('receiverPhone') && !invalidFields.includes('receiverPhone') && !isValidNumber(formatPhoneNumber(address.receiverPhone, 'ZA'), 'ZA')) {\n    invalidFields.push('receiverPhone');\n    $('#bash--input-receiverPhone').addClass('invalid');\n    $('#bash--label-receiverPhone').focus();\n  }\n  return {\n    isValid: !invalidFields.length,\n    invalidFields: invalidFields\n  };\n};\nvar setCartClasses = function setCartClasses() {\n  var items = window.vtexjs.checkout.orderForm.items;\n  var _getSpecialCategories2 = (0,_utils_functions__WEBPACK_IMPORTED_MODULE_1__.getSpecialCategories)(items),\n    hasTVs = _getSpecialCategories2.hasTVs,\n    hasSimCards = _getSpecialCategories2.hasSimCards,\n    hasFurnitureMixed = _getSpecialCategories2.hasFurnitureMixed;\n  var $container = '#shipping-data';\n  if (hasTVs) {\n    $(\"\".concat($container, \":not(.has-tv)\")).addClass('has-tv');\n  } else {\n    $(\"\".concat($container, \".has-tv\")).removeClass('has-tv');\n  }\n  if (hasSimCards) {\n    $(\"\".concat($container, \":not(.has-rica)\")).addClass('has-rica');\n  } else {\n    $(\"\".concat($container, \".has-rica\")).removeClass('has-rica');\n  }\n  if (hasFurnitureMixed) {\n    $(\"\".concat($container, \":not(.has-furniture-mixed)\")).addClass('has-furniture-mixed');\n  } else {\n    $(\"\".concat($container, \".has-furniture-mixed\")).removeClass('has-furniture-mixed');\n  }\n};\nvar updateDeliveryFeeDisplay = function updateDeliveryFeeDisplay() {\n  var _window$vtexjs$checko;\n  if (!window.vtexjs.checkout.orderForm.totalizers) return;\n  var _ref3 = window.vtexjs.checkout.orderForm.totalizers.find(function (item) {\n      return item.id === 'Shipping';\n    }) || {\n      value: ((_window$vtexjs$checko = window.vtexjs.checkout.orderForm) === null || _window$vtexjs$checko === void 0 || (_window$vtexjs$checko = _window$vtexjs$checko.shippingData) === null || _window$vtexjs$checko === void 0 || (_window$vtexjs$checko = _window$vtexjs$checko.address) === null || _window$vtexjs$checko === void 0 ? void 0 : _window$vtexjs$checko.addressType) === _utils_const__WEBPACK_IMPORTED_MODULE_0__.AD_TYPE.PICKUP ? _utils_const__WEBPACK_IMPORTED_MODULE_0__.COLLECT_FEE : _utils_const__WEBPACK_IMPORTED_MODULE_0__.DELIVERY_FEE\n    },\n    shippingFee = _ref3.value;\n  var feeText = 'Free';\n  if (shippingFee > 0) feeText = \"R\".concat((shippingFee / 100).toFixed(2).replace('.00', ''));\n  if ($('#bash--delivery-fee').length > 0) {\n    document.getElementById('bash--delivery-fee').innerHTML = feeText;\n  }\n};\nvar customShippingDataIsValid = function customShippingDataIsValid() {\n  var _window$vtexjs$checko2;\n  var items = (_window$vtexjs$checko2 = window.vtexjs.checkout.orderForm) === null || _window$vtexjs$checko2 === void 0 ? void 0 : _window$vtexjs$checko2.items;\n  var _getSpecialCategories3 = (0,_utils_functions__WEBPACK_IMPORTED_MODULE_1__.getSpecialCategories)(items),\n    hasTVs = _getSpecialCategories3.hasTVs,\n    hasSimCards = _getSpecialCategories3.hasSimCards;\n  var valid = true;\n  if (hasTVs) {\n    var data = (0,_utils_services__WEBPACK_IMPORTED_MODULE_4__.getOrderFormCustomData)(_utils_const__WEBPACK_IMPORTED_MODULE_0__.TV_APP);\n    if (!data.tvID) valid = false;\n  }\n  if (hasSimCards) {\n    var _data = (0,_utils_services__WEBPACK_IMPORTED_MODULE_4__.getOrderFormCustomData)(_utils_const__WEBPACK_IMPORTED_MODULE_0__.RICA_APP);\n    if (!_data.idOrPassport || !_data.streetAddress || !_data.postalCode) valid = false;\n  }\n  return valid;\n};\nvar populateDeliveryError = function populateDeliveryError() {\n  var errors = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : [];\n  if ($('#bash-delivery-error-container').length < 1) return;\n  var errorsHtml = errors.length > 0 ? errors.map(function (error) {\n    return (0,_DeliveryError__WEBPACK_IMPORTED_MODULE_5__.DeliveryError)(error);\n  }) : '';\n  $('#bash-delivery-error-container').html(errorsHtml);\n  if (errors.length > 0) $('html, body').animate({\n    scrollTop: $('#bash-delivery-error-container').offset().top\n  }, 400);\n};\nvar showAlertBox = function showAlertBox() {\n  var alertText = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : 'Address saved';\n  $('.alert-container').addClass('show');\n  $('.alert-container').slideDown();\n  $('#bash-alert-container').html((0,_Elements_Alert__WEBPACK_IMPORTED_MODULE_6__.Alert)({\n    text: alertText\n  }));\n  // After 5 seconds, remove the element\n  setTimeout(function () {\n    $('.alert-container').slideUp();\n  }, 5000);\n};\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (mapGoogleAddress);\n\n//# sourceURL=webpack://custom-shipping-step-by-items/./src/partials/Deliver/utils.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"addressIsValid\": () => (/* binding */ addressIsValid),\n/* harmony export */   \"clearRicaFields\": () => (/* binding */ clearRicaFields),\n/* harmony export */   \"customShippingDataIsValid\": () => (/* binding */ customShippingDataIsValid),\n/* harmony export */   \"getBestRecipient\": () => (/* binding */ getBestRecipient),\n/* harmony export */   \"parseAttribute\": () => (/* binding */ parseAttribute),\n/* harmony export */   \"populateAddressForm\": () => (/* binding */ populateAddressForm),\n/* harmony export */   \"populateDeliveryError\": () => (/* binding */ populateDeliveryError),\n/* harmony export */   \"populateExtraFields\": () => (/* binding */ populateExtraFields),\n/* harmony export */   \"populateRicaFields\": () => (/* binding */ populateRicaFields),\n/* harmony export */   \"populateTVFields\": () => (/* binding */ populateTVFields),\n/* harmony export */   \"provinceShortCode\": () => (/* binding */ provinceShortCode),\n/* harmony export */   \"setCartClasses\": () => (/* binding */ setCartClasses),\n/* harmony export */   \"setDeliveryLoading\": () => (/* binding */ setDeliveryLoading),\n/* harmony export */   \"setPickupLoading\": () => (/* binding */ setPickupLoading),\n/* harmony export */   \"showAlertBox\": () => (/* binding */ showAlertBox),\n/* harmony export */   \"updateDeliveryFeeDisplay\": () => (/* binding */ updateDeliveryFeeDisplay)\n/* harmony export */ });\n/* harmony import */ var _utils_const__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../../utils/const */ \"./src/utils/const.js\");\n/* harmony import */ var _utils_functions__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../../utils/functions */ \"./src/utils/functions.js\");\n/* harmony import */ var _utils_isInSouthAfrica__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../../utils/isInSouthAfrica */ \"./src/utils/isInSouthAfrica.js\");\n/* harmony import */ var _utils_phoneFields__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ../../utils/phoneFields */ \"./src/utils/phoneFields.js\");\n/* harmony import */ var _utils_services__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ../../utils/services */ \"./src/utils/services.js\");\n/* harmony import */ var _DeliveryError__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! ./DeliveryError */ \"./src/partials/Deliver/DeliveryError.js\");\n/* harmony import */ var _Elements_Alert__WEBPACK_IMPORTED_MODULE_6__ = __webpack_require__(/*! ./Elements/Alert */ \"./src/partials/Deliver/Elements/Alert.js\");\n/* harmony import */ var _constants__WEBPACK_IMPORTED_MODULE_7__ = __webpack_require__(/*! ./constants */ \"./src/partials/Deliver/constants.js\");\n/* harmony import */ var _utils_phoneNumberFormat__WEBPACK_IMPORTED_MODULE_8__ = __webpack_require__(/*! ../../utils/phoneNumberFormat */ \"./src/utils/phoneNumberFormat.js\");\nfunction _typeof(obj) { \"@babel/helpers - typeof\"; return _typeof = \"function\" == typeof Symbol && \"symbol\" == typeof Symbol.iterator ? function (obj) { return typeof obj; } : function (obj) { return obj && \"function\" == typeof Symbol && obj.constructor === Symbol && obj !== Symbol.prototype ? \"symbol\" : typeof obj; }, _typeof(obj); }\nfunction _toConsumableArray(arr) { return _arrayWithoutHoles(arr) || _iterableToArray(arr) || _unsupportedIterableToArray(arr) || _nonIterableSpread(); }\nfunction _nonIterableSpread() { throw new TypeError(\"Invalid attempt to spread non-iterable instance.\\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.\"); }\nfunction _iterableToArray(iter) { if (typeof Symbol !== \"undefined\" && iter[Symbol.iterator] != null || iter[\"@@iterator\"] != null) return Array.from(iter); }\nfunction _arrayWithoutHoles(arr) { if (Array.isArray(arr)) return _arrayLikeToArray(arr); }\nfunction _regeneratorRuntime() { \"use strict\"; /*! regenerator-runtime -- Copyright (c) 2014-present, Facebook, Inc. -- license (MIT): https://github.com/facebook/regenerator/blob/main/LICENSE */ _regeneratorRuntime = function _regeneratorRuntime() { return exports; }; var exports = {}, Op = Object.prototype, hasOwn = Op.hasOwnProperty, defineProperty = Object.defineProperty || function (obj, key, desc) { obj[key] = desc.value; }, $Symbol = \"function\" == typeof Symbol ? Symbol : {}, iteratorSymbol = $Symbol.iterator || \"@@iterator\", asyncIteratorSymbol = $Symbol.asyncIterator || \"@@asyncIterator\", toStringTagSymbol = $Symbol.toStringTag || \"@@toStringTag\"; function define(obj, key, value) { return Object.defineProperty(obj, key, { value: value, enumerable: !0, configurable: !0, writable: !0 }), obj[key]; } try { define({}, \"\"); } catch (err) { define = function define(obj, key, value) { return obj[key] = value; }; } function wrap(innerFn, outerFn, self, tryLocsList) { var protoGenerator = outerFn && outerFn.prototype instanceof Generator ? outerFn : Generator, generator = Object.create(protoGenerator.prototype), context = new Context(tryLocsList || []); return defineProperty(generator, \"_invoke\", { value: makeInvokeMethod(innerFn, self, context) }), generator; } function tryCatch(fn, obj, arg) { try { return { type: \"normal\", arg: fn.call(obj, arg) }; } catch (err) { return { type: \"throw\", arg: err }; } } exports.wrap = wrap; var ContinueSentinel = {}; function Generator() {} function GeneratorFunction() {} function GeneratorFunctionPrototype() {} var IteratorPrototype = {}; define(IteratorPrototype, iteratorSymbol, function () { return this; }); var getProto = Object.getPrototypeOf, NativeIteratorPrototype = getProto && getProto(getProto(values([]))); NativeIteratorPrototype && NativeIteratorPrototype !== Op && hasOwn.call(NativeIteratorPrototype, iteratorSymbol) && (IteratorPrototype = NativeIteratorPrototype); var Gp = GeneratorFunctionPrototype.prototype = Generator.prototype = Object.create(IteratorPrototype); function defineIteratorMethods(prototype) { [\"next\", \"throw\", \"return\"].forEach(function (method) { define(prototype, method, function (arg) { return this._invoke(method, arg); }); }); } function AsyncIterator(generator, PromiseImpl) { function invoke(method, arg, resolve, reject) { var record = tryCatch(generator[method], generator, arg); if (\"throw\" !== record.type) { var result = record.arg, value = result.value; return value && \"object\" == _typeof(value) && hasOwn.call(value, \"__await\") ? PromiseImpl.resolve(value.__await).then(function (value) { invoke(\"next\", value, resolve, reject); }, function (err) { invoke(\"throw\", err, resolve, reject); }) : PromiseImpl.resolve(value).then(function (unwrapped) { result.value = unwrapped, resolve(result); }, function (error) { return invoke(\"throw\", error, resolve, reject); }); } reject(record.arg); } var previousPromise; defineProperty(this, \"_invoke\", { value: function value(method, arg) { function callInvokeWithMethodAndArg() { return new PromiseImpl(function (resolve, reject) { invoke(method, arg, resolve, reject); }); } return previousPromise = previousPromise ? previousPromise.then(callInvokeWithMethodAndArg, callInvokeWithMethodAndArg) : callInvokeWithMethodAndArg(); } }); } function makeInvokeMethod(innerFn, self, context) { var state = \"suspendedStart\"; return function (method, arg) { if (\"executing\" === state) throw new Error(\"Generator is already running\"); if (\"completed\" === state) { if (\"throw\" === method) throw arg; return { value: void 0, done: !0 }; } for (context.method = method, context.arg = arg;;) { var delegate = context.delegate; if (delegate) { var delegateResult = maybeInvokeDelegate(delegate, context); if (delegateResult) { if (delegateResult === ContinueSentinel) continue; return delegateResult; } } if (\"next\" === context.method) context.sent = context._sent = context.arg;else if (\"throw\" === context.method) { if (\"suspendedStart\" === state) throw state = \"completed\", context.arg; context.dispatchException(context.arg); } else \"return\" === context.method && context.abrupt(\"return\", context.arg); state = \"executing\"; var record = tryCatch(innerFn, self, context); if (\"normal\" === record.type) { if (state = context.done ? \"completed\" : \"suspendedYield\", record.arg === ContinueSentinel) continue; return { value: record.arg, done: context.done }; } \"throw\" === record.type && (state = \"completed\", context.method = \"throw\", context.arg = record.arg); } }; } function maybeInvokeDelegate(delegate, context) { var methodName = context.method, method = delegate.iterator[methodName]; if (undefined === method) return context.delegate = null, \"throw\" === methodName && delegate.iterator[\"return\"] && (context.method = \"return\", context.arg = undefined, maybeInvokeDelegate(delegate, context), \"throw\" === context.method) || \"return\" !== methodName && (context.method = \"throw\", context.arg = new TypeError(\"The iterator does not provide a '\" + methodName + \"' method\")), ContinueSentinel; var record = tryCatch(method, delegate.iterator, context.arg); if (\"throw\" === record.type) return context.method = \"throw\", context.arg = record.arg, context.delegate = null, ContinueSentinel; var info = record.arg; return info ? info.done ? (context[delegate.resultName] = info.value, context.next = delegate.nextLoc, \"return\" !== context.method && (context.method = \"next\", context.arg = undefined), context.delegate = null, ContinueSentinel) : info : (context.method = \"throw\", context.arg = new TypeError(\"iterator result is not an object\"), context.delegate = null, ContinueSentinel); } function pushTryEntry(locs) { var entry = { tryLoc: locs[0] }; 1 in locs && (entry.catchLoc = locs[1]), 2 in locs && (entry.finallyLoc = locs[2], entry.afterLoc = locs[3]), this.tryEntries.push(entry); } function resetTryEntry(entry) { var record = entry.completion || {}; record.type = \"normal\", delete record.arg, entry.completion = record; } function Context(tryLocsList) { this.tryEntries = [{ tryLoc: \"root\" }], tryLocsList.forEach(pushTryEntry, this), this.reset(!0); } function values(iterable) { if (iterable || \"\" === iterable) { var iteratorMethod = iterable[iteratorSymbol]; if (iteratorMethod) return iteratorMethod.call(iterable); if (\"function\" == typeof iterable.next) return iterable; if (!isNaN(iterable.length)) { var i = -1, next = function next() { for (; ++i < iterable.length;) if (hasOwn.call(iterable, i)) return next.value = iterable[i], next.done = !1, next; return next.value = undefined, next.done = !0, next; }; return next.next = next; } } throw new TypeError(_typeof(iterable) + \" is not iterable\"); } return GeneratorFunction.prototype = GeneratorFunctionPrototype, defineProperty(Gp, \"constructor\", { value: GeneratorFunctionPrototype, configurable: !0 }), defineProperty(GeneratorFunctionPrototype, \"constructor\", { value: GeneratorFunction, configurable: !0 }), GeneratorFunction.displayName = define(GeneratorFunctionPrototype, toStringTagSymbol, \"GeneratorFunction\"), exports.isGeneratorFunction = function (genFun) { var ctor = \"function\" == typeof genFun && genFun.constructor; return !!ctor && (ctor === GeneratorFunction || \"GeneratorFunction\" === (ctor.displayName || ctor.name)); }, exports.mark = function (genFun) { return Object.setPrototypeOf ? Object.setPrototypeOf(genFun, GeneratorFunctionPrototype) : (genFun.__proto__ = GeneratorFunctionPrototype, define(genFun, toStringTagSymbol, \"GeneratorFunction\")), genFun.prototype = Object.create(Gp), genFun; }, exports.awrap = function (arg) { return { __await: arg }; }, defineIteratorMethods(AsyncIterator.prototype), define(AsyncIterator.prototype, asyncIteratorSymbol, function () { return this; }), exports.AsyncIterator = AsyncIterator, exports.async = function (innerFn, outerFn, self, tryLocsList, PromiseImpl) { void 0 === PromiseImpl && (PromiseImpl = Promise); var iter = new AsyncIterator(wrap(innerFn, outerFn, self, tryLocsList), PromiseImpl); return exports.isGeneratorFunction(outerFn) ? iter : iter.next().then(function (result) { return result.done ? result.value : iter.next(); }); }, defineIteratorMethods(Gp), define(Gp, toStringTagSymbol, \"Generator\"), define(Gp, iteratorSymbol, function () { return this; }), define(Gp, \"toString\", function () { return \"[object Generator]\"; }), exports.keys = function (val) { var object = Object(val), keys = []; for (var key in object) keys.push(key); return keys.reverse(), function next() { for (; keys.length;) { var key = keys.pop(); if (key in object) return next.value = key, next.done = !1, next; } return next.done = !0, next; }; }, exports.values = values, Context.prototype = { constructor: Context, reset: function reset(skipTempReset) { if (this.prev = 0, this.next = 0, this.sent = this._sent = undefined, this.done = !1, this.delegate = null, this.method = \"next\", this.arg = undefined, this.tryEntries.forEach(resetTryEntry), !skipTempReset) for (var name in this) \"t\" === name.charAt(0) && hasOwn.call(this, name) && !isNaN(+name.slice(1)) && (this[name] = undefined); }, stop: function stop() { this.done = !0; var rootRecord = this.tryEntries[0].completion; if (\"throw\" === rootRecord.type) throw rootRecord.arg; return this.rval; }, dispatchException: function dispatchException(exception) { if (this.done) throw exception; var context = this; function handle(loc, caught) { return record.type = \"throw\", record.arg = exception, context.next = loc, caught && (context.method = \"next\", context.arg = undefined), !!caught; } for (var i = this.tryEntries.length - 1; i >= 0; --i) { var entry = this.tryEntries[i], record = entry.completion; if (\"root\" === entry.tryLoc) return handle(\"end\"); if (entry.tryLoc <= this.prev) { var hasCatch = hasOwn.call(entry, \"catchLoc\"), hasFinally = hasOwn.call(entry, \"finallyLoc\"); if (hasCatch && hasFinally) { if (this.prev < entry.catchLoc) return handle(entry.catchLoc, !0); if (this.prev < entry.finallyLoc) return handle(entry.finallyLoc); } else if (hasCatch) { if (this.prev < entry.catchLoc) return handle(entry.catchLoc, !0); } else { if (!hasFinally) throw new Error(\"try statement without catch or finally\"); if (this.prev < entry.finallyLoc) return handle(entry.finallyLoc); } } } }, abrupt: function abrupt(type, arg) { for (var i = this.tryEntries.length - 1; i >= 0; --i) { var entry = this.tryEntries[i]; if (entry.tryLoc <= this.prev && hasOwn.call(entry, \"finallyLoc\") && this.prev < entry.finallyLoc) { var finallyEntry = entry; break; } } finallyEntry && (\"break\" === type || \"continue\" === type) && finallyEntry.tryLoc <= arg && arg <= finallyEntry.finallyLoc && (finallyEntry = null); var record = finallyEntry ? finallyEntry.completion : {}; return record.type = type, record.arg = arg, finallyEntry ? (this.method = \"next\", this.next = finallyEntry.finallyLoc, ContinueSentinel) : this.complete(record); }, complete: function complete(record, afterLoc) { if (\"throw\" === record.type) throw record.arg; return \"break\" === record.type || \"continue\" === record.type ? this.next = record.arg : \"return\" === record.type ? (this.rval = this.arg = record.arg, this.method = \"return\", this.next = \"end\") : \"normal\" === record.type && afterLoc && (this.next = afterLoc), ContinueSentinel; }, finish: function finish(finallyLoc) { for (var i = this.tryEntries.length - 1; i >= 0; --i) { var entry = this.tryEntries[i]; if (entry.finallyLoc === finallyLoc) return this.complete(entry.completion, entry.afterLoc), resetTryEntry(entry), ContinueSentinel; } }, \"catch\": function _catch(tryLoc) { for (var i = this.tryEntries.length - 1; i >= 0; --i) { var entry = this.tryEntries[i]; if (entry.tryLoc === tryLoc) { var record = entry.completion; if (\"throw\" === record.type) { var thrown = record.arg; resetTryEntry(entry); } return thrown; } } throw new Error(\"illegal catch attempt\"); }, delegateYield: function delegateYield(iterable, resultName, nextLoc) { return this.delegate = { iterator: values(iterable), resultName: resultName, nextLoc: nextLoc }, \"next\" === this.method && (this.arg = undefined), ContinueSentinel; } }, exports; }\nfunction asyncGeneratorStep(gen, resolve, reject, _next, _throw, key, arg) { try { var info = gen[key](arg); var value = info.value; } catch (error) { reject(error); return; } if (info.done) { resolve(value); } else { Promise.resolve(value).then(_next, _throw); } }\nfunction _asyncToGenerator(fn) { return function () { var self = this, args = arguments; return new Promise(function (resolve, reject) { var gen = fn.apply(self, args); function _next(value) { asyncGeneratorStep(gen, resolve, reject, _next, _throw, \"next\", value); } function _throw(err) { asyncGeneratorStep(gen, resolve, reject, _next, _throw, \"throw\", err); } _next(undefined); }); }; }\nfunction _slicedToArray(arr, i) { return _arrayWithHoles(arr) || _iterableToArrayLimit(arr, i) || _unsupportedIterableToArray(arr, i) || _nonIterableRest(); }\nfunction _nonIterableRest() { throw new TypeError(\"Invalid attempt to destructure non-iterable instance.\\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.\"); }\nfunction _unsupportedIterableToArray(o, minLen) { if (!o) return; if (typeof o === \"string\") return _arrayLikeToArray(o, minLen); var n = Object.prototype.toString.call(o).slice(8, -1); if (n === \"Object\" && o.constructor) n = o.constructor.name; if (n === \"Map\" || n === \"Set\") return Array.from(o); if (n === \"Arguments\" || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)) return _arrayLikeToArray(o, minLen); }\nfunction _arrayLikeToArray(arr, len) { if (len == null || len > arr.length) len = arr.length; for (var i = 0, arr2 = new Array(len); i < len; i++) arr2[i] = arr[i]; return arr2; }\nfunction _iterableToArrayLimit(arr, i) { var _i = null == arr ? null : \"undefined\" != typeof Symbol && arr[Symbol.iterator] || arr[\"@@iterator\"]; if (null != _i) { var _s, _e, _x, _r, _arr = [], _n = !0, _d = !1; try { if (_x = (_i = _i.call(arr)).next, 0 === i) { if (Object(_i) !== _i) return; _n = !1; } else for (; !(_n = (_s = _x.call(_i)).done) && (_arr.push(_s.value), _arr.length !== i); _n = !0); } catch (err) { _d = !0, _e = err; } finally { try { if (!_n && null != _i[\"return\"] && (_r = _i[\"return\"](), Object(_r) !== _r)) return; } finally { if (_d) throw _e; } } return _arr; } }\nfunction _arrayWithHoles(arr) { if (Array.isArray(arr)) return arr; }\n// @ts-nocheck\n\n\n\n\n\n\n\n\n\nvar _usePhoneNumberFormat = (0,_utils_phoneNumberFormat__WEBPACK_IMPORTED_MODULE_8__[\"default\"])(),\n  formatPhoneNumber = _usePhoneNumberFormat.formatPhoneNumber,\n  isValidNumber = _usePhoneNumberFormat.isValidNumber;\nvar setDeliveryLoading = function setDeliveryLoading() {\n  var _document$querySelect;\n  (_document$querySelect = document.querySelector('.bash--delivery-container')) === null || _document$querySelect === void 0 ? void 0 : _document$querySelect.classList.add('shimmer');\n};\nvar setPickupLoading = function setPickupLoading() {\n  var _document$querySelect2, _document$querySelect3;\n  (_document$querySelect2 = document.querySelector('.delivery-group-content')) === null || _document$querySelect2 === void 0 || (_document$querySelect2 = _document$querySelect2.classList) === null || _document$querySelect2 === void 0 ? void 0 : _document$querySelect2.add('shimmer');\n  (_document$querySelect3 = document.querySelector('.vtex-omnishipping-1-x-ask')) === null || _document$querySelect3 === void 0 || (_document$querySelect3 = _document$querySelect3.classList) === null || _document$querySelect3 === void 0 ? void 0 : _document$querySelect3.add('shimmer');\n};\nvar provinceShortCode = function provinceShortCode(province) {\n  switch (province) {\n    case 'Select':\n      return '';\n    case 'Western Cape':\n      return 'WC';\n    case 'Easter Cape':\n      return 'EC';\n    case 'Gauteng':\n      return 'GP';\n    case 'KwaZulu-Natal':\n    case 'KwaZulu Natal':\n      return 'KZN';\n    case 'Northern Cape':\n      return 'NC';\n    case 'Limpopo':\n      return 'LP';\n    case 'Mpumalanga':\n      return 'MP';\n    case 'North West':\n      return 'NW';\n    case 'Freestate':\n    case 'Free State':\n      return 'FS';\n    default:\n      return province;\n  }\n};\nvar getBestRecipient = function getBestRecipient(_ref) {\n  var _window, _window2, _window3, _document$getElementB;\n  var _ref$preferred = _ref.preferred,\n    preferred = _ref$preferred === void 0 ? undefined : _ref$preferred,\n    _ref$type = _ref.type,\n    type = _ref$type === void 0 ? 'delivery' : _ref$type;\n  var firstName = (_window = window) === null || _window === void 0 || (_window = _window.vtexjs) === null || _window === void 0 || (_window = _window.checkout) === null || _window === void 0 || (_window = _window.orderForm) === null || _window === void 0 || (_window = _window.clientProfileData) === null || _window === void 0 ? void 0 : _window.firstName;\n  var lastName = (_window2 = window) === null || _window2 === void 0 || (_window2 = _window2.vtexjs) === null || _window2 === void 0 || (_window2 = _window2.checkout) === null || _window2 === void 0 || (_window2 = _window2.orderForm) === null || _window2 === void 0 || (_window2 = _window2.clientProfileData) === null || _window2 === void 0 ? void 0 : _window2.lastName;\n  var shippingReceiverName = (_window3 = window) === null || _window3 === void 0 || (_window3 = _window3.vtexjs) === null || _window3 === void 0 || (_window3 = _window3.checkout) === null || _window3 === void 0 || (_window3 = _window3.orderForm) === null || _window3 === void 0 || (_window3 = _window3.shippingData) === null || _window3 === void 0 || (_window3 = _window3.address) === null || _window3 === void 0 ? void 0 : _window3.receiverName;\n  var clientProfileName = \"\".concat(firstName !== null && firstName !== void 0 ? firstName : '', \" \").concat(lastName !== null && lastName !== void 0 ? lastName : '').trim();\n  if (type === 'collect') return preferred || shippingReceiverName || clientProfileName || '';\n  return preferred || ((_document$getElementB = document.getElementById('client-first-name')) === null || _document$getElementB === void 0 ? void 0 : _document$getElementB.value) || clientProfileName || '';\n};\nvar populateAddressForm = function populateAddressForm(address) {\n  var _document$getElementB2, _document$getElementB3;\n  var number = address.number,\n    street = address.street,\n    addressType = address.addressType,\n    businessName = address.businessName,\n    companyBuilding = address.companyBuilding,\n    neighborhood = address.neighborhood,\n    postalCode = address.postalCode,\n    state = address.state,\n    city = address.city,\n    receiverName = address.receiverName,\n    receiverPhone = address.receiverPhone,\n    complement = address.complement,\n    id = address.id,\n    addressId = address.addressId,\n    addressName = address.addressName,\n    geoCoordinate = address.geoCoordinate;\n\n  // Clear any populated fields\n  (_document$getElementB2 = document.getElementById('bash--address-form')) === null || _document$getElementB2 === void 0 ? void 0 : _document$getElementB2.reset();\n  (0,_utils_functions__WEBPACK_IMPORTED_MODULE_1__.hideBusinessName)();\n  var lat;\n  var lng;\n  try {\n    var _JSON$parse = JSON.parse(JSON.stringify(geoCoordinate));\n    var _JSON$parse2 = _slicedToArray(_JSON$parse, 2);\n    lng = _JSON$parse2[0];\n    lat = _JSON$parse2[1];\n    if (!(0,_utils_isInSouthAfrica__WEBPACK_IMPORTED_MODULE_2__[\"default\"])([lng, lat])) {\n      // Coordinates are most likely still at [lat, lng]\n      if (_utils_isInSouthAfrica__WEBPACK_IMPORTED_MODULE_2__[\"default\"][(lat, lng)]) {\n        lng = lat;\n        lat = lng;\n      } else {\n        // Unset the coordinates\n        lng = 0;\n        lat = 0;\n      }\n    }\n  } catch (e) {\n    console.warn('Could not parse geo coords', {\n      address: address,\n      geoCoordinate: geoCoordinate\n    });\n  }\n\n  // Only overwrite defaults if values exist.\n  if (receiverName) document.getElementById('bash--input-receiverName').value = receiverName !== null && receiverName !== void 0 ? receiverName : '';\n  if (complement) document.getElementById('bash--input-complement').value = complement !== null && complement !== void 0 ? complement : '';\n\n  // addressId indicates that address is being edited / completed.\n  if (id || addressId) document.getElementById('bash--input-addressId').value = id || addressId; // TODO remove this?\n  if (addressName) document.getElementById('bash--input-addressName').value = addressName;\n  var streetLine = \"\".concat(number ? \"\".concat(number, \" \") : '').concat(street).replace(\", \".concat(companyBuilding), '');\n\n  // Address type\n  if (addressType === 'commercial') {\n    $('#radio-addressType-business').click();\n    (0,_utils_functions__WEBPACK_IMPORTED_MODULE_1__.showBusinessName)({\n      focus: false\n    });\n  }\n  if (businessName) document.getElementById('bash--input-businessName').value = businessName;\n  document.getElementById('bash--input-number').value = '';\n  document.getElementById('bash--input-street').value = streetLine || '';\n  document.getElementById('bash--input-companyBuilding').value = companyBuilding || '';\n  document.getElementById('bash--input-neighborhood').value = neighborhood || '';\n  document.getElementById('bash--input-city').value = city || '';\n  document.getElementById('bash--input-postalCode').value = postalCode || '';\n  document.getElementById('bash--input-state').value = provinceShortCode(state);\n  document.getElementById('bash--input-lat').value = lat || '';\n  document.getElementById('bash--input-lng').value = lng || '';\n  var fields = (0,_utils_services__WEBPACK_IMPORTED_MODULE_4__.getOrderFormCustomData)(_utils_const__WEBPACK_IMPORTED_MODULE_0__.PICKUP);\n\n  // Only overwrite defaults if values exist.\n  if (receiverName) document.getElementById('bash--input-receiverName').value = receiverName !== null && receiverName !== void 0 ? receiverName : '';\n  if (complement) document.getElementById('bash--input-complement').value = complement !== null && complement !== void 0 ? complement : '';\n  document.getElementById('bash--input-receiverPhone').value = formatPhoneNumber(receiverPhone).trim() || (0,_utils_phoneFields__WEBPACK_IMPORTED_MODULE_3__.getBestPhoneNumber)({\n    preferred: receiverPhone,\n    type: 'delivery',\n    fields: fields\n  });\n  (_document$getElementB3 = document.getElementById('bash--input-receiverPhone')) === null || _document$getElementB3 === void 0 ? void 0 : _document$getElementB3.addEventListener('onblur', function (event) {\n    event.target.value = event.target.value.trim();\n  });\n  $(':invalid').trigger('change');\n};\nvar parseAttribute = function parseAttribute(data) {\n  try {\n    return JSON.parse(decodeURIComponent(data));\n  } catch (e) {\n    return undefined;\n  }\n};\nvar populateExtraFields = function populateExtraFields(address, fields) {\n  var prefix = arguments.length > 2 && arguments[2] !== undefined ? arguments[2] : '';\n  var override = arguments.length > 3 && arguments[3] !== undefined ? arguments[3] : false;\n  if (!address) return;\n  for (var i = 0; i < fields.length; i++) {\n    var fieldId = \"bash--input-\".concat(prefix).concat(fields[i]);\n    if (document.getElementById(fieldId) && (address[fields[i]] || override) && (!document.getElementById(fieldId).value || override)) {\n      document.getElementById(fieldId).value = address[fields[i]];\n    }\n  }\n  $(':invalid').trigger('change');\n};\nvar populateRicaFields = function populateRicaFields() {\n  var _document$getElementB4;\n  var address = window.vtexjs.checkout.orderForm.shippingData.address;\n  if ((_document$getElementB4 = document.getElementById('bash--input-rica_streetAddress')) !== null && _document$getElementB4 !== void 0 && _document$getElementB4.value || !address) return;\n  address.fullName = getBestRecipient({\n    type: 'delivery'\n  });\n  address.streetAddress = address.street;\n  address.suburb = address.neighborhood;\n  address.province = address.state;\n  populateExtraFields(address, _constants__WEBPACK_IMPORTED_MODULE_7__.requiredRicaFields, 'rica_');\n  var data = (0,_utils_services__WEBPACK_IMPORTED_MODULE_4__.getOrderFormCustomData)(_utils_const__WEBPACK_IMPORTED_MODULE_0__.RICA_APP);\n  if (data.streetAddress) populateExtraFields(data, _constants__WEBPACK_IMPORTED_MODULE_7__.requiredRicaFields, 'rica_', true);\n};\nvar clearRicaFields = function clearRicaFields() {\n  var idOrPassport = $('#bash--input-rica_idOrPassport').val();\n  var clearedRica = {\n    idOrPassport: idOrPassport !== null && idOrPassport !== void 0 ? idOrPassport : '',\n    // TODO populate with users ID\n    fullName: '',\n    streetAddress: '',\n    suburb: '',\n    city: '',\n    postalCode: '',\n    province: ''\n  };\n  populateExtraFields(clearedRica, _constants__WEBPACK_IMPORTED_MODULE_7__.requiredRicaFields, 'rica_', true);\n};\nvar populateTVFields = /*#__PURE__*/function () {\n  var _ref2 = _asyncToGenerator( /*#__PURE__*/_regeneratorRuntime().mark(function _callee() {\n    var data;\n    return _regeneratorRuntime().wrap(function _callee$(_context) {\n      while (1) switch (_context.prev = _context.next) {\n        case 0:\n          data = (0,_utils_services__WEBPACK_IMPORTED_MODULE_4__.getOrderFormCustomData)(_utils_const__WEBPACK_IMPORTED_MODULE_0__.TV_APP);\n          populateExtraFields(data, _constants__WEBPACK_IMPORTED_MODULE_7__.requiredTVFields, 'tv');\n        case 2:\n        case \"end\":\n          return _context.stop();\n      }\n    }, _callee);\n  }));\n  return function populateTVFields() {\n    return _ref2.apply(this, arguments);\n  };\n}();\n\n// Runs when you setAddress\nvar addressIsValid = function addressIsValid(address) {\n  var validateExtraFields = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : true;\n  var items = window.vtexjs.checkout.orderForm.items;\n  var _getSpecialCategories = (0,_utils_functions__WEBPACK_IMPORTED_MODULE_1__.getSpecialCategories)(items),\n    hasTVs = _getSpecialCategories.hasTVs,\n    hasSimCards = _getSpecialCategories.hasSimCards;\n  var requiredFields = _constants__WEBPACK_IMPORTED_MODULE_7__.requiredAddressFields;\n  var invalidFields = [];\n  if (hasTVs && validateExtraFields) {\n    requiredFields = [].concat(_toConsumableArray(requiredFields), _toConsumableArray(_constants__WEBPACK_IMPORTED_MODULE_7__.requiredTVFields));\n  }\n  if (hasSimCards && validateExtraFields) {\n    requiredFields = [].concat(_toConsumableArray(requiredFields), _toConsumableArray(_constants__WEBPACK_IMPORTED_MODULE_7__.requiredRicaFields));\n  }\n  for (var i = 0; i < requiredFields.length; i++) {\n    if (!address[requiredFields[i]]) invalidFields.push(requiredFields[i]);\n  }\n  if (requiredFields.includes('receiverPhone') && !invalidFields.includes('receiverPhone') && !isValidNumber(formatPhoneNumber(address.receiverPhone, 'ZA'), 'ZA')) {\n    invalidFields.push('receiverPhone');\n    $('#bash--input-receiverPhone').addClass('invalid');\n    $('#bash--label-receiverPhone').focus();\n  }\n  return {\n    isValid: !invalidFields.length,\n    invalidFields: invalidFields\n  };\n};\nvar setCartClasses = function setCartClasses() {\n  var items = window.vtexjs.checkout.orderForm.items;\n  var _getSpecialCategories2 = (0,_utils_functions__WEBPACK_IMPORTED_MODULE_1__.getSpecialCategories)(items),\n    hasTVs = _getSpecialCategories2.hasTVs,\n    hasSimCards = _getSpecialCategories2.hasSimCards,\n    hasFurnitureMixed = _getSpecialCategories2.hasFurnitureMixed;\n  var $container = '#shipping-data';\n  if (hasTVs) {\n    $(\"\".concat($container, \":not(.has-tv)\")).addClass('has-tv');\n  } else {\n    $(\"\".concat($container, \".has-tv\")).removeClass('has-tv');\n  }\n  if (hasSimCards) {\n    $(\"\".concat($container, \":not(.has-rica)\")).addClass('has-rica');\n  } else {\n    $(\"\".concat($container, \".has-rica\")).removeClass('has-rica');\n  }\n  if (hasFurnitureMixed) {\n    $(\"\".concat($container, \":not(.has-furniture-mixed)\")).addClass('has-furniture-mixed');\n  } else {\n    $(\"\".concat($container, \".has-furniture-mixed\")).removeClass('has-furniture-mixed');\n  }\n};\nvar updateDeliveryFeeDisplay = function updateDeliveryFeeDisplay() {\n  var _window$vtexjs$checko;\n  if (!window.vtexjs.checkout.orderForm.totalizers) return;\n  var _ref3 = window.vtexjs.checkout.orderForm.totalizers.find(function (item) {\n      return item.id === 'Shipping';\n    }) || {\n      value: ((_window$vtexjs$checko = window.vtexjs.checkout.orderForm) === null || _window$vtexjs$checko === void 0 || (_window$vtexjs$checko = _window$vtexjs$checko.shippingData) === null || _window$vtexjs$checko === void 0 || (_window$vtexjs$checko = _window$vtexjs$checko.address) === null || _window$vtexjs$checko === void 0 ? void 0 : _window$vtexjs$checko.addressType) === _utils_const__WEBPACK_IMPORTED_MODULE_0__.AD_TYPE.PICKUP ? _utils_const__WEBPACK_IMPORTED_MODULE_0__.COLLECT_FEE : _utils_const__WEBPACK_IMPORTED_MODULE_0__.DELIVERY_FEE\n    },\n    shippingFee = _ref3.value;\n  var feeText = 'Free';\n  if (shippingFee > 0) feeText = \"R\".concat((shippingFee / 100).toFixed(2).replace('.00', ''));\n  if ($('#bash--delivery-fee').length > 0) {\n    document.getElementById('bash--delivery-fee').innerHTML = feeText;\n  }\n};\nvar customShippingDataIsValid = function customShippingDataIsValid() {\n  var _window$vtexjs$checko2;\n  var items = (_window$vtexjs$checko2 = window.vtexjs.checkout.orderForm) === null || _window$vtexjs$checko2 === void 0 ? void 0 : _window$vtexjs$checko2.items;\n  var _getSpecialCategories3 = (0,_utils_functions__WEBPACK_IMPORTED_MODULE_1__.getSpecialCategories)(items),\n    hasTVs = _getSpecialCategories3.hasTVs,\n    hasSimCards = _getSpecialCategories3.hasSimCards;\n  var valid = true;\n  if (hasTVs) {\n    var data = (0,_utils_services__WEBPACK_IMPORTED_MODULE_4__.getOrderFormCustomData)(_utils_const__WEBPACK_IMPORTED_MODULE_0__.TV_APP);\n    if (!data.tvID) valid = false;\n  }\n  if (hasSimCards) {\n    var _data = (0,_utils_services__WEBPACK_IMPORTED_MODULE_4__.getOrderFormCustomData)(_utils_const__WEBPACK_IMPORTED_MODULE_0__.RICA_APP);\n    if (!_data.idOrPassport || !_data.streetAddress || !_data.postalCode) valid = false;\n  }\n  return valid;\n};\nvar populateDeliveryError = function populateDeliveryError() {\n  var errors = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : [];\n  if ($('#bash-delivery-error-container').length < 1) return;\n  var errorsHtml = errors.length > 0 ? errors.map(function (error) {\n    return (0,_DeliveryError__WEBPACK_IMPORTED_MODULE_5__.DeliveryError)(error);\n  }) : '';\n  $('#bash-delivery-error-container').html(errorsHtml);\n  if (errors.length > 0) $('html, body').animate({\n    scrollTop: $('#bash-delivery-error-container').offset().top\n  }, 400);\n};\nvar showAlertBox = function showAlertBox() {\n  var alertText = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : 'Address saved';\n  $('.alert-container').addClass('show');\n  $('.alert-container').slideDown();\n  $('#bash-alert-container').html((0,_Elements_Alert__WEBPACK_IMPORTED_MODULE_6__.Alert)({\n    text: alertText\n  }));\n  // After 5 seconds, remove the element\n  setTimeout(function () {\n    $('.alert-container').slideUp();\n  }, 5000);\n};\n\n//# sourceURL=webpack://custom-shipping-step-by-items/./src/partials/Deliver/utils.js?");
 
 /***/ }),
